% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_tbl.R
\name{agg_tbl}
\alias{agg_tbl}
\title{Aggregates a data.table}
\usage{
agg_tbl(dt, metric, metric_name, dimension, dimension_name, agg_fn,
  na.rm = TRUE)
}
\arguments{
\item{dt}{A data.table object}

\item{metric}{A vector of numeric variable(s) to aggregate}

\item{metric_name}{A vector of variable name(s) for the numeric variable(s) to aggregate. Defaults to metric.}

\item{dimension}{A vector of dimension variable(s) to aggregate by}

\item{dimension_name}{A vector of dimension variable name(s) for the dimension variable(s). Defaults to dimension.}

\item{agg_fn}{An aggregation function}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
Returns a data.table
}
\description{
Aggregates a data.table
}
\examples{
\dontrun{
to follow
}
}

